layout/generic/test/test_bug394239.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=394239
-->
<head>
  <title>Test for Bug 394239</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style>
   #a::after { content:"anonymous text"; }
  </style>
</head>
<body>
<div id="content">
 <object id="a" style="position: relative;">
  <div style="position: absolute;">
   <input>
  </div>&#1593;
  <div style="position: absolute;">
    <span style="position: fixed;">t</span>
  </div>
 </object>
</div>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=394239">Mozilla Bug 394239</a>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.expectAssertions(1);

/** Test for Bug 394239 **/
function test()
{
  $("a").style.direction="rtl";
  ok(true, "Should not crash");
  SimpleTest.finish();
}
  setTimeout(test, 500);
  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
